bot_fakturama/requirements.txt
botcity-framework-core>=1.2.1,<2.0
botcity-maestro-sdk>=0.3.3,<1.0

botcity-excel-plugin
